multi_agents_example/requirements.txt
langchain>=1.0.0
langchain-openai>=1.0.0
langchain_community>=0.4.0
langgraph>=0.2.0
SQLAlchemy==2.0.32
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
